🔧 chore - structure the project
