+++ main.py
@@ -1,15 +1,14 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-import os
+from app.routers import auth, upload_pdf, ask_question, get_presentation
 from dotenv import load_dotenv
+import os
 
 # Load environment variables from .env file only in local development
 if os.getenv("CI") is None:
     # Set the ENVIRONMENT variable to "development"
     os.environ["ENVIRONMENT"] = "development"
     load_dotenv()
-
-from app.api.routes import router as api_router
 
 app = FastAPI()
 
@@ -21,4 +20,7 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-app.include_router(api_router, prefix="/api/v1")
+app.include_router(auth.router, prefix="/api/v1", tags=["Authentication"])
+app.include_router(upload_pdf.router, prefix="/api/v1", tags=["Upload PDF"])
+app.include_router(ask_question.router, prefix="/api/v1", tags=["Ask Question"])
+app.include_router(get_presentation.router, prefix="/api/v1", tags=["Generate Presentation"])
