+++ tests/test_4_get_presentation.py
@@ -4,7 +4,7 @@ import pytest
 import os
 from main import app
 import time
-from app.api.knowledgebase_handler import get_knowledge_base, set_knowledge_base
+from app.service.knowledgebase_handler import get_knowledge_base, set_knowledge_base
 
 client = TestClient(app)
